slimjim can now autodetect MAC and IP of the victim
It also supports operation on an 802.1x network, passing
EAPOL packets across the bridge to the real authenticator.

It doesn't attempt to choose an upstream MAC/IP to use
to connect to the victim, however. That should best be
done on a case by case basis, under manual control.

There is now a pcap-based script that monitors ARP and
other IP traffic to identify link local data, non-local
link traffic, DNS name to IP mappings, as well as finding
the nameservers in use.

Link local data is required to talk to other hosts on the
same subnet, since we cannot ARP, routes to other subnets
make it possible to identify local routers, and which other
subnets exist. DNS responses allow us to obtain names of
systems being talked to, as well as make connections to
those systems without having to look up the names to IP
address mappings ourself. And finally, identifying addresses
of name servers potentially allows for DNS tunnelled C2
connections.