Add virtual interface to the namespace
This allows for direct access into the namespace for things
like SSH (and by extension, WireShark over SSH, and forwarded
channels, etc). Starts an SSH daemon by default on the virtual
interface.

This can be accessed by using the public IP address as a jump
host:

ssh -J nanopi-r1 root@169.254.20.1

Note that the IP address chosen by default is a link-local
range, hopefully that should not conflict with any addresses
actively in use on the target network.